% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCTA_env_cor.R
\name{res_cor}
\alias{res_cor}
\title{Calculate the residual correlation from a fitted bivariate GREML model}
\usage{
res_cor(Ve1, Ve2, Ce)
}
\arguments{
\item{Ve1}{A numeric value. The residual variance of trait 1.}

\item{Ve2}{A numeric value. The residual variance of trait 2.}

\item{Ce}{A numeric value. The residual covariance between
trait 1 and 2.}
}
\value{
Residual correlation between trait 1 and 2
}
\description{
\code{write_gpf_syntax} Bivariate GREML as implemented in GCTA outputs
genetic correlation but not residual correlation. This function calculates
the residual correlation based on the variances and covariances of both
traits.
}
\details{
Formula from Trzaskowski, M., Yang, J., Visscher, P. M., & Plomin, R. (2013).
DNA evidence for strong genetic stability and increasing heritability
of intelligence from age 7 to 12. Molecular psychiatry, 19(3), 380-384.
}

