% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCTA_env_cor.R
\name{res_se}
\alias{res_se}
\title{Calculate the standard error for the
residual correlation from a fitted bivariate GREML model}
\usage{
res_se(Ve1, VarVe1, Ve2, VarVe2, Ce, VarCe, CovVe1Ve2, CovVe1Ce, CovVe2Ce)
}
\arguments{
\item{Ve1}{A numeric value. The residual variance of trait 1.}

\item{VarVe1}{A numeric value. Sampling variance of trait 1.
Can be found in 1,1 of covariance matrix.}

\item{Ve2}{A numeric value. The residual variance of trait 2.}

\item{VarVe2}{A numeric value.
The sampling variance of trait 2.
Can be found in 2,2 of covariance matrix.}

\item{Ce}{A numeric value. The residual covariance between
trait 1 and 2.}

\item{VarCe}{A numeric value. Sampling Variance of Ce [square(SE of C(e)_tr2)].
Can be found in 3,3 of covariance matrix.}

\item{CovVe1Ve2}{A numeric value. Sampling covariance between Ve1 and Ve2.
Can be found in 5,4 in Covariance Matrix}

\item{CovVe1Ce}{A numeric value. Sampling covariance between Ve1 and Ce.
Can be found in 6,4 in Covariance Matrix.}

\item{CovVe2Ce}{A numeric value. Sampling covariance between Ve2 and Ce.
Can be found in 6,5 in Covariance Matrix}
}
\value{
Standard error of residual correlation between trait 1 and 2
}
\description{
\code{write_gpf_syntax} Bivariate GREML as implemented in GCTA outputs
genetic correlation but not residual correlation. This function calculates
the standard error of the residual correlation based on the variances and
covariances of both traits. See covariance matrix of GCTA output.
}
\details{
Formula from Trzaskowski, M., Yang, J., Visscher, P. M., & Plomin, R. (2013).
DNA evidence for strong genetic stability and increasing heritability
of intelligence from age 7 to 12. Molecular psychiatry, 19(3), 380-384.
}

